Created newItem function using jQuery
Basic tasks for app:
The user can add a new item to a list of items.
The user can cross out an item from the list of items.
The user can delete an item from the list of items.
The user can change the order of items in the list of items.